pyproject.toml
[tool.poetry]
name = "jigsaw_toxic_classification_ml"
version = "0.1.0"
description = "Professional, reproducible ML project for Jigsaw Toxic Comment Classification (multi-label)."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "jigsaw_toxic_classification_ml" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^2.1.0"
pandas = "^2.2.0"
pyarrow = "^18.0.0"
scikit-learn = "^1.5.0"
hydra-core = "^1.3.2"
pydantic = "^2.8.2"
matplotlib = "^3.9.0"
mlflow = "^2.16.0"
onnx = "^1.16.2"
onnxruntime = "^1.19.2"
dvc = "^3.55.2"
fire = "^0.6.0"
rich = "^13.8.1"

[tool.poetry.group.train.dependencies]
torch = "^2.4.0"
pytorch-lightning = "^2.4.0"
torchmetrics = "^1.4.1"
onnxscript = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.2"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["jigsaw_toxic_classification_ml"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
